Generating Random Exams from 'LaTeX' documents

This package is designed with exams and homework created in 'LaTeX' in mind. It allows to randomize and personalize exams and homework and it aids the user with grading them.

Details

If you are using the exam class from 'LaTeX' already, it is likely that this program works as it is.

If you just want to randomize your exams,

Look at vignette("BasicUse", package = "TexExamRandomizer") for an introduction of the concept behind this library and a quick way to start using it.

Look at vignette("ExamOptions", package = "TexExamRandomizer") for a more detailed explanations of what options can be used on a document.

If instead you are trying to use the library to create your own randomizer for a certain use you might have, you should start by looking at CreateRandomExams and GenerateHomework.

See also

Author

Alejandro Gonzalez Recuenco

e-mail: alejandrogonzalezrecuenco@gmail.com